Adds cache
