+++ app/src/main/java/com/mzazi/harrypotterandroid/data/cache/model/CharacterEntity.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mzazi.harrypotterandroid.data.cache
+package com.mzazi.harrypotterandroid.data.cache.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
